raftstore: extract batch system (#6414)
