+++ Cargo.toml
@@ -137,6 +137,7 @@ engine = { path = "components/engine" }
 tikv_util = { path = "components/tikv_util" }
 farmhash = "1.1.5"
 external_storage = { path = "components/external_storage" }
+batch-system = { path = "components/batch-system" }
 
 [dependencies.murmur3]
 git = "https://github.com/pingcap/murmur3.git"
@@ -196,6 +197,7 @@ members = [
   "components/external_storage",
   "components/backup",
   "cmd",
+  "components/batch-system",
 ]
 
 [profile.dev]
