+++ components/batch-system/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "batch-system"
+version = "0.1.0"
+edition = "2018"
+
+[dependencies]
+crossbeam = "0.7"
+tikv_util = { path = "../tikv_util" }
+slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
+slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "91904ade" }
+
+[dev-dependencies]
+derive_more = "0.15.0"
+
+[[test]]
+name = "tests"
+path = "tests/cases/mod.rs"
