+++ components/batch-system/tests/cases/batch.rs
@@ -1,0 +1,43 @@
+// Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
+
+use crate::*;
+use batch_system::*;
+use std::time::Duration;
+use tikv_util::mpsc;
+
+#[test]
+fn test_batch() {
+    let (control_tx, control_fsm) = new_runner(10);
+    let (router, mut system) = batch_system::create_system(2, 2, control_tx, control_fsm);
+    let builder = Builder::new();
+    let metrics = builder.metrics.clone();
+    system.spawn("test".to_owned(), builder);
+    let mut expected_metrics = HandleMetrics::default();
+    assert_eq!(*metrics.lock().unwrap(), expected_metrics);
+    let (tx, rx) = mpsc::unbounded();
+    let tx_ = tx.clone();
+    let r = router.clone();
+    router
+        .send_control(Some(Box::new(move |_: &mut Runner| {
+            let (tx, runner) = new_runner(10);
+            let mailbox = BasicMailbox::new(tx, runner);
+            r.register(1, mailbox);
+            tx_.send(1).unwrap();
+        })))
+        .unwrap();
+    assert_eq!(rx.recv_timeout(Duration::from_secs(3)), Ok(1));
+    router
+        .send(
+            1,
+            Some(Box::new(move |_: &mut Runner| {
+                tx.send(2).unwrap();
+            })),
+        )
+        .unwrap();
+    assert_eq!(rx.recv_timeout(Duration::from_secs(3)), Ok(2));
+    system.shutdown();
+    expected_metrics.control = 1;
+    expected_metrics.normal = 1;
+    expected_metrics.begin = 2;
+    assert_eq!(*metrics.lock().unwrap(), expected_metrics);
+}
