+++ src/raftstore/store/fsm/mod.rs
@@ -5,10 +5,8 @@
 //! stores. They are mixed for now, will be separated in the future.
 
 pub mod apply;
-mod batch;
 mod metrics;
 mod peer;
-mod router;
 pub mod store;
 
 pub use self::apply::{
@@ -17,11 +15,7 @@ pub use self::apply::{
     Msg as ApplyTask, Notifier as ApplyNotifier, Proposal, RegionProposal, Registration,
     TaskRes as ApplyTaskRes,
 };
-pub use self::batch::{
-    BatchRouter, BatchSystem, Fsm, HandlerBuilder, NormalScheduler, PollHandler,
-};
-pub use self::peer::{DestroyPeerJob, GroupState};
-pub use self::router::{BasicMailbox, Mailbox};
+pub use self::peer::{DestroyPeerJob, GroupState, PeerFsm};
 pub use self::store::{
     create_raft_batch_system, new_compaction_listener, RaftBatchSystem, RaftPollerBuilder,
     RaftRouter, StoreInfo,
